1) 1) displays.c/h: rename DDC_IO_Mode to MCCS_IO_Mode, add value USB_IO to the enum.  Change the name ddc_io_mode to io_mode in struct Display_Ref. The remaining changes are the effect of this change throughout the code.
In many cases, code of the form
   if io_mode == DDC_IO_DEVI2C) {
      ...
   }
   else {    // assume DDC_IO_ADL
   ...
   }

has been replaced by a switch statement which is less prone to silently
failing if an additional IO mode is added.

In many places, dummy code with messages is has been inserted for
case USB_IO.

Note that in a lot of the lower levels of the code, the USB_IO option makes
no sense.