change DEFAULT_EDID_READ_SIZE to 0 (dynamic)
In release 1.0.0, the EDID read size was increased from 128 to 256 bytes for
greater compatibility, though ddcutil uses only the first 128 bytes.

This read size can cause reading the EDID to fail for displays connected using
a docking station based on the DisplayPort i915 standard, and using the i915
driver in kernels since 5.11, which finally implemented I2C support on docking
stations.  Command line option --edid-read-size=128 was implemented as a workaround
until the bug is fixed in the i915 driver.

This change causes i2c_get_edid_by_fd() to instead try both 128 and 256 byte
reads of the EDID.