is_phantom_display(): fix edid comparison
Just comparing 128 byte EDIDs to see if they are for the same
display can fail.  As seen in user reports, the EDIDs for the
valid and phantom displays can differ in the display type bits
of byte 24, with one indicating RGB 4:4:4 and the other RGB 4:4:4 +
YCrCb 4:4:2.  To make the variation even more secure, these bits
are set only for EDID 1.4.

This change instead calls new function edid_ids_match(), which
compares th manufacturer, model id (binary and chararacter) and
serial number (binary and character) to determine if the EDIDs match.