options --enable/disable-try-get-edid-from-sysfs replace utility option --f14
Controls whether bus examination attempts to obtain the EDID from /sys
instead of reading it itself from the monitor.  This is faster, and not
subject to cross-instance locks, but there appear to be situations
where a stale EDID is presented even after a monitor has been disconnected
or even replaced.

Also --i1 specifies the number of seconds stabilized_connector_names()
waits if a disconnected display was detected.