ddci_init(): submaster_initializer not called if implicit initialization
ddca_init()/ddca_init2() can be called "explicitly" or "implicitly".
An explicit call is when they ares called directly by the client program.
If an API function dependent on initialization is called before
explicit initialization, ddca_init() is called with arguments such
that it can never fail.  This is an implicit call.  In particular,
it is called with option DDCA_INIT_OPTIONS_DISABLE_CONFIG_FILE and
a null libopts string in order to make failure impossible.

The bug was that some initialization was skipped with this combination
of ddca_init() args. In particular, submaster_initializer(), which among
other things sets the value of sys_drm_connectors, was not called.
Also any options in the ddcutilrc configuration file that turned on
tracing were not processed, making debugging more difficult.

PowerDevil calls API functions requiring initialization before calling
ddca_init(). As a result, there's an implicit call to ddca_init(),
and because of the bug sys_drm_connectors was not set, resulting
in the "assert(sys_drm_connectors)" failure in function i2c_check_bus().