ddca_get_display_refs(), ddca_get_display_info_list2() always return 0. Addresses issue #417.
Errors that occur opening individual displays or reading their EDIDs
are still reported using ddca_get_error_detail().
In addition, error messages are written to the terminal and, depending
on the current syslog level, to the system log.

API documentation in ddcutil_c_api.h is updated to reflect this behavior.